Separation of concerns & menu looping
